+++ src/main.js
@@ -1,5 +1,6 @@
 import Vue from 'vue'
 import App from './App.vue'
+import y from './page/y.vue'
 
 new Vue({
   el: '#app',
